refactor: Moved Cpu Information Module to Individual OS-Specific files
