+++ src/sysinfo_linux.c
@@ -1,3 +1,4 @@
+#include "sysinfo.h"
 #include <stdio.h>
 #include <string.h>
 #include <time.h>
@@ -20,4 +21,27 @@ FOUND:
   fclose(f);
   sscanf(buf + 5, "%ld", &boot_time);
   return boot_time;
+}
+
+int sys_cpu_info(struct cpu_info_t *curr) {
+  FILE *f;
+  char buf[256];
+
+  f = fopen("/proc/stat", "r");
+  if (!f) {
+    perror("fopen /proc/stat");
+    return -1;
+  }
+
+  while (fgets(buf, sizeof buf, f)) {
+    if (!strncmp(buf, "cpu ", 4)) {
+      int n = sscanf(buf + 5, "%llu %llu %llu %llu", &curr->u_mode, &curr->nice,
+                     &curr->s_mode, &curr->idle);
+      fclose(f);
+      return (n == 4) ? 0 : -1;
+    }
+  }
+
+  fclose(f);
+  return -1;
 }
